refactor: fixture 패키지 분리
